FBP-294. Added XSDs and bindings
