+++ activiti-intellij-plugin/src/main/kotlin/com/valb3r/bpmn/intellij/activiti/plugin/schemas/ActivitiBpmnFileSchemasProvider.kt
@@ -1,25 +1,10 @@
 package com.valb3r.bpmn.intellij.activiti.plugin.schemas
 
-import com.intellij.openapi.module.Module
-import com.intellij.openapi.vfs.VfsUtil
-import com.intellij.psi.PsiFile
-import com.intellij.psi.xml.XmlFile
-import com.intellij.xml.XmlSchemaProvider
-import com.valb3r.bpmn.intellij.plugin.core.settings.currentSettings
+import com.valb3r.bpmn.intellij.plugin.core.schemas.BpmnFileSchemasProvider
 
 
-class ActivitiBpmnFileSchemasProvider: XmlSchemaProvider() {
+class ActivitiBpmnFileSchemasProvider: BpmnFileSchemasProvider() {
 
-    private val BPMN20Schema = "/xsds/BPMN20.xsd"
-
-    override fun isAvailable(file: XmlFile): Boolean {
-        return currentSettings().openExtensions.any { file.name.endsWith(it) }
-    }
-
-    override fun getSchema(url: String, module: Module?, baseFile: PsiFile): XmlFile? {
-        val resource = ActivitiBpmnFileSchemasProvider::class.java.getResource(BPMN20Schema) ?: return null
-        val fileByURL = VfsUtil.findFileByURL(resource) ?: return null
-        val result = baseFile.manager.findFile(fileByURL)
-        return if (result is XmlFile) { result.copy() as XmlFile } else null
-    }
+    override val BPMN20Schemas: Map<String, String>
+        get() = super.BPMN20Schemas + ("http://activiti.org/bpmn" to "/xsds/activiti-6.0.xsd")
 }
