+++ flowable-intellij-plugin/src/main/resources/META-INF/plugin.xml
@@ -304,6 +304,7 @@
     <extensions defaultExtensionNs="com.intellij">
         <fileType name="Flowable BPMN 2.0 Engine XML process" implementationClass="com.valb3r.bpmn.intellij.plugin.flowable.meta.FlowableFileType"
                   fieldName="INSTANCE" language="XML" patterns="*.bpmn20.xml"/>
+        <xml.schemaProvider implementation="com.valb3r.bpmn.intellij.plugin.flowable.schemas.FlowableBpmnFileSchemasProvider"/>
     </extensions>
     <!-- File templates registration -->
     <extensions defaultExtensionNs="com.intellij">
